[deploy] Forward installations to installationhandler instead. Fixes JB#48107
The PackageKit install methods should be limited to only privileged egid callers, so sdk-deploy-rpm needs to call the sailfish-installationhandler daemon with its install requests so it can check sideloading restrictions and prompt the user.

As this process will wait for a success/failure signal from the installationhandler service, it also watches for the service dying so it doesn't get stuck.